S3 Content-Type fix.
Boto overwrites the Header Content-Type setting with its own guess.  Changed _save to add the Content-Type to the metadata.
The metadata is added to the Header after boto overwrites the Content-Type in the Header.

In the case of uploaded files, the file's content_type is ignored and instead mimetypes.guess_type is used instead.
This fix checks for content.content_type and if present, uses it instead of mimetypes.guess_type.

We needed this fix to properly set the Content-Type when uploading caf files without an extension.